fix for yellow soul shots crashing
the yellow soul shots try calling audio_sound_gain on the return value of snd_play_ex (which is undefined, because it never returns anything) this causes it to crash. Other parts of the code use this function in the same way, and probably crash in a similar way.

This was crashing on my version of the decomp, main branch, no changes.